requirements.txt
torch>=1.9.0
transformers>=4.5.0
networkx>=2.6.0
jieba>=0.42.1
numpy>=1.19.5
scikit-learn>=0.24.2
streamlit>=1.0.0
pyvis>=0.1.9
requests>=2.26.0
pandas>=1.5.3
rdflib>=6.3.2
py2neo>=2021.2.3
beautifulsoup4>=4.11.2
tushare>=1.2.89
neo4j>=5.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
pandas>=1.3.0 